# apps/core-frontend/config/swagger-config/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "authControllerLogin",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginDto" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthMessageResponse" }
              }
            }
          },
          "401": { "description": "Invalid credentials" }
        },
        "summary": "Login with email and password",
        "tags": ["auth"]
      }
    },
    "/api/auth/register": {
      "post": {
        "operationId": "authControllerRegister",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/RegisterDto" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration successful",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/AuthMessageResponse" }
              }
            }
          }
        },
        "summary": "Register a new account",
        "tags": ["auth"]
      }
    },
...
  },
  "info": {
    "title": "TrustMed Core API",
    "description": "API documentation for the TrustMed Core backend",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "user@example.com" },
          "password": { "type": "string", "example": "password123" }
        },
        "required": ["email", "password"]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "example": "user@example.com" },
          "password": { "type": "string", "example": "password123" },
          "firstName": { "type": "string", "example": "John" },
          "lastName": { "type": "string", "example": "Doe" }
        },
        "required": ["email", "password", "firstName"]
      },
      "AuthMessageResponse": {
        "type": "object",
        "properties": {
          "message": { "type": "string", "example": "Login successful" }
        },
        "required": ["message"]
      },
...
    }
  }
}